refactor: consolidate keychain and device auth into a local-auth domain
