+++ src/components/expanded-state/DevTestBackupState.tsx
@@ -7,8 +7,8 @@ import { resetInternetCredentials } from 'react-native-keychain';
 import CopyTooltip from '@/components/copy-tooltip';
 import { Box, Text } from '@/design-system';
 import { createBackup, restoreBackup } from '@/features/backup/backup';
+import { clear } from '@/features/local-auth/keychain';
 import { WalletLoadingStates } from '@/helpers/walletLoadingStates';
-import { clear } from '@/keychain';
 import { logger, RainbowError } from '@/logger';
 import { clearAllStorages } from '@/model/mmkv';
 import Navigation from '@/navigation/Navigation';
