+++ src/components/expanded-state/profile/ProfileModal.tsx
@@ -2,6 +2,7 @@ import React, { useCallback, useRef } from 'react';
 import { Platform, View } from 'react-native';
 
 import Divider from '@/components/Divider';
+import BiometricButtonContent from '@/features/local-auth/components/BiometricButtonContent';
 import styled from '@/framework/ui/styled-thing';
 import { opacity } from '@/framework/ui/utils/opacity';
 import * as i18n from '@/languages';
@@ -9,7 +10,6 @@ import { margin, padding, position } from '@/styles';
 import { useTheme } from '@/theme/ThemeContext';
 
 import ButtonPressAnimation from '../../animations/ButtonPressAnimation';
-import { BiometricButtonContent } from '../../buttons';
 import CopyTooltip from '../../copy-tooltip';
 import { Centered, ColumnWithDividers } from '../../layout';
 import { AvatarCircle } from '../../profile';
