+++ src/components/hold-to-activate-button/HoldToActivateButton.tsx
@@ -16,11 +16,11 @@ import { triggerHaptics } from 'react-native-turbo-haptics';
 import { AnimatedSpinner } from '@/components/animations/AnimatedSpinner';
 import { SPRING_CONFIGS } from '@/components/animations/animationConfigs';
 import ShimmerAnimation from '@/components/animations/ShimmerAnimation';
-import { useBiometryIconString } from '@/components/buttons/BiometricButtonContent';
 import HoldToAuthorizeButtonIcon from '@/components/buttons/hold-to-authorize/HoldToAuthorizeButtonIcon';
 import { LedgerIcon } from '@/components/icons/svg/LedgerIcon';
 import { Box, Text, useColorMode, type BoxProps, type TextProps } from '@/design-system';
 import { getColorForTheme } from '@/design-system/color/useForegroundColor';
+import { useBiometryIconString } from '@/features/local-auth/components/BiometricButtonContent';
 import { time } from '@/framework/core/utils/time';
 import { useWalletsStore } from '@/state/wallets/walletsStore';
 import { colors } from '@/styles';
