+++ src/helpers/index.ts
@@ -1,4 +1,3 @@
-export { default as BiometryTypes } from './biometryTypes';
 export { default as CurrencySelectionTypes } from './currencySelectionTypes';
 export { default as isKeyboardOpen } from './isKeyboardOpen';
 export { default as isReanimatedAvailable } from './isReanimatedAvailable';
