+++ src/screens/CheckIdentifierScreen.tsx
@@ -10,8 +10,8 @@ import ButtonPressAnimation from '@/components/animations/ButtonPressAnimation';
 import { ImgixImage } from '@/components/images';
 import { SimpleSheet } from '@/components/sheet/SimpleSheet';
 import { BackgroundProvider, Bleed, Box, Inline, Inset, Separator, Stack, Text } from '@/design-system';
+import * as kc from '@/features/local-auth/keychain';
 import walletBackupStepTypes from '@/helpers/walletBackupStepTypes';
-import * as kc from '@/keychain';
 import * as i18n from '@/languages';
 import { logger, RainbowError } from '@/logger';
 import { getHeightForStep } from '@/navigation/config';
