+++ src/screens/Diagnostics/index.tsx
@@ -9,17 +9,17 @@ import { getDeviceId } from '@/analytics/utils';
 import { SimpleSheet } from '@/components/sheet/SimpleSheet';
 import { Toast, ToastPositionContainer } from '@/components/toasts';
 import { BackgroundProvider, Box } from '@/design-system';
+import { privateKeyKey, seedPhraseKey } from '@/features/local-auth/keychainConstants';
+import { loadAllKeys } from '@/features/local-auth/legacyKeychain';
+import { authenticateWithPINAndCreateIfNeeded } from '@/features/local-auth/pinAuthentication';
 import AesEncryptor from '@/handlers/aesEncryption';
-import { authenticateWithPINAndCreateIfNeeded } from '@/handlers/authentication';
 import useWalletsWithBalancesAndNames from '@/hooks/useWalletsWithBalancesAndNames';
 import * as i18n from '@/languages';
 import { logger, RainbowError } from '@/logger';
-import { loadAllKeys } from '@/model/keychain';
 import { useNavigation } from '@/navigation/Navigation';
 import Routes from '@/navigation/routesNames';
 import { type RootStackParamList } from '@/navigation/types';
 import { DiagnosticsContent } from '@/screens/Diagnostics/DiagnosticsContent';
-import { privateKeyKey, seedPhraseKey } from '@/utils/keychainConstants';
 import { deriveAccountFromWalletInput } from '@/utils/wallet';
 
 import { createAndShareStateDumpFile } from './helpers/createAndShareStateDumpFile';
