+++ src/screens/token-launcher/components/TokenLauncherFooter.tsx
@@ -12,11 +12,11 @@ import RainbowCoinIcon from '@/components/coin-icon/RainbowCoinIcon';
 import { HoldToActivateButton } from '@/components/hold-to-activate-button/HoldToActivateButton';
 import { Box, Inline, Text } from '@/design-system';
 import { getColorForTheme } from '@/design-system/color/useForegroundColor';
+import useBiometryType from '@/features/local-auth/hooks/useBiometryType';
+import * as kc from '@/features/local-auth/keychain';
+import BiometryTypes from '@/features/local-auth/types/biometryTypes';
 import { buildTokenDeeplink } from '@/handlers/deeplinks';
-import { BiometryTypes } from '@/helpers';
-import useBiometryType from '@/hooks/useBiometryType';
 import { useTokenLauncher } from '@/hooks/useTokenLauncher';
-import * as kc from '@/keychain';
 import * as i18n from '@/languages';
 import { ensureError, logger, RainbowError } from '@/logger';
 import { loadPrivateKey } from '@/model/wallet';
